Compute the effect coefficients from the DLL, and stop shipping them
Ports the coefficient decode from DotNetAdministravit. The reverb and chorus
numbers were held to exist nowhere in the file -- computed by the engine at
start-up, so obtaining them meant executing SCCore.dll on Windows x64 -- which
is why assets/presets.json was committed and compiled into the library, and
why NOTICE.md carried it as an exception to the no-redistribution rule.

They are in the file, encoded: per-character rows of ring taps stored relative
to a base the loader adds (fx_load_dsp_preset, 0x2000 for every GS macro), and
coefficient rows in a signed-14-bit float encoding (fixed14_to_float).
Searching for tap 9118 finds nothing because the stored value is 926. The
register program supplies the rest -- injection tap from the pre-delay byte,
damping from the pre-LPF ladder, tank feedback from the reverb time, chorus
tap geometry from the macro bytes.

Two things C# gave for free need care here: the wide shift in tap_base relies
on the original's 32-bit register wrap, and the mantissa shift in fixed14 is
signed, so both run through uint32_t rather than invoking undefined behaviour.
Getting either wrong yields plausible-but-wrong coefficients, not a crash.

EffectPresets::ensure_from runs the decode from NoteRenderer's constructor, so
opening the engine is what makes the effects available. defaults() hands out a
reference, so installing a set retires the previous one rather than destroying
it. A pinned harvest via use() or TABULASONORA_PRESETS still wins.

Verified field for field against the retired presets.json: all eight reverb
macros, all eight chorus macros and both defaults, no residual. The bit-exact
gates are unmoved and a WATRWLD1 render is unchanged at peak 0.312144. Two
callers reached for presets without ever opening a ROM and only worked by
accident of ordering: the effects tests now ask for them explicitly, and the
CLI's dump-effect requires the DLL like every other subcommand.

+++ tests/tone_generator_tests.cpp
@@ -607,9 +607,9 @@ TEST_CASE("the NRPN-dropped crash still sounds on the SC-55 map", "[stream][scco
 
         int count = 0;
         for (std::size_t i = 1; i < left.size(); ++i) {
-            const double previous = left[i - 1] + right[i - 1];
-            const double current = left[i] + right[i];
-            if ((previous < 0.0) != (current < 0.0)) {
+            const float previous = left[i - 1] + right[i - 1];
+            const float current = left[i] + right[i];
+            if ((previous < 0.0F) != (current < 0.0F)) {
                 ++count;
             }
         }
